Rename file to index.ts
